bypass potential gradle bug
